Changing TypeStatus to combobox with PickList change the default of comboboxes to textpl Added color pref for Synon in trees Changed Info Icon Moved SpecifyAppPrefs (out of test)
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@4904 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/conversion/GenericDBConversion.java
@@ -2532,7 +2532,7 @@ public class GenericDBConversion implements IdMapperIndexIncrementerIFace
                 pl.setSizeLimit(-1);
                 pl.setIsSystem(true);
                 pl.setTableName("preptype");
-                pl.setType(1);
+                pl.setType((byte)1);
                 
             } else
             {
